Better access restriction implemented
Sorry for the mess commit, but changing all the const stuff was messy
and I (stupidly) didn't start from a clean state.

Members that should be const are const.
        => It's better but because of this, we can't use an array in out
        LastCellFinder and need to use a vector instead (from what I
        gather). I think it's worth it.
Added a convenience method to get a region based on an int
Fix a missing ';' at the end of a Visitor class